Relocate 'ErrorBoundary' component unit test folders
